Fix image annotation/label git references
Same purpose as the previously reverted commit.  However this commit
keeps the script's $REPO_URL pointing at the upstream source.  This is
required so that the underlying `build-push.sh` script can correctly
derrive the destination image registry.  This problem could not have
been discovered by the test-build CI job in a PR.

Signed-off-by: Chris Evich <cevich@redhat.com>