Fix podman use of remote git content
Container context directories should be self-contained, but the various
podman flavors were not.  All of them referred to a file on github that
no longer exists.  For now, duplicate the necessary files and add them
directly.  A future commit may de-duplicate the three flavors
(stable/testing/upstream) into a single Containerfile (similar to how
it's done in buildah).

Signed-off-by: Chris Evich <cevich@redhat.com>