Deduplicate podman & skopeo build context
There are three flavors of the podman|buildah|skopeo images.  For
`podman` and `skopeo`, previously each flavor had a subdirectory with
largely duplicate `Containerfile` and other content.  Fix this situation
by utilizing a `flavor` build augment, in the same manner as done by the
buildah image.  Update the relevant CI/CD jobs to cope with the shared
content.

Signed-off-by: Chris Evich <cevich@redhat.com>