Support rebuildable 'stable' flavor containers
Currently, container images are point-in-time builds.  Whatever
happens to be present in the package repositories at build-time is
content included in the images.  This is undesirable from a user
perspective as it results in the image contents frequently changing.
It's also undesirable from an operational perspective as there is no
way to recover a prior image should it become corrupt or get pruned
from the registry server.

Assist a future migration to a static package list by introducing
a script to help produce package lists for each architecture.  This
script is intended to be run periodically by automation, with the
output being checked into version-control assuming test-builds pass
CI.

Signed-off-by: Chris Evich <cevich@redhat.com>