Force amd64 arch for temporary build
The current and prior CI VMs contain cached copies of the latest Fedora
container image for each arch.  Unfortunately the last arch pulled is
`arm64`, which results in it being the default image used by podman when
`--arch` is not specified.  Ensure this can never cause problems by
hard-coding the `amd64` arch which will always match the VM.

Signed-off-by: Chris Evich <cevich@redhat.com>