Force amd64 arch for temporary build
With CI VMs `c20240320t153921z` and earlier, they contain cached copies
of the latest Fedora container image for each arch.  Unfortunately the
last arch pulled is `arm64`, which results in it being the default image
used by podman when `--arch` is not specified.  Until/if this is fixed in
the VM image, hard-code an `amd64` arch which will always match the VM.

Signed-off-by: Chris Evich <cevich@redhat.com>