Make buildah image smaller w/ ucpp vs cpp
It's unlikely that a full-blown C pre-pocessor is required.  My
understanding is this package was added in order to support the
`Containerfile.in` `#include` feature.  This can easily be accomplished
with `ucpp` at a greatly reduced overall image size.  Realize this
effect while de-duplicating the package list via a build-arg.

Signed-off-by: Chris Evich <cevich@redhat.com>